Changed to SoftwareSerial
Regular Serial.begin(BAUD) was causing way negative temp readings.  Works with SoftwareSerial library connected to same pins, however, everytime you connect to the serial port, it seems to reboot the MicroView.  Not sure if this is the loader board causing the reset, or part of the library.  Works well enough for what I need.