Updating the .ino firmware
Took the station on a "mobile" run which involved an altitude drop of around 1000 ft.  At the lower level (where pressure is higher) the firmware was stuck in the while loop trap that prevents poor readings from the barometer.  I broke this into 2.  The trap was to prevent a negative value that occasionally happens, but I put an upper limit thinking it would be good.  Decided it is okay to have a higher limit but wanted to give it a retry to eventually break out and take other measurements.